Adds a change indicator denotating latency changes.
Adds a change indicator to the end of the log line denotating if we've
seen a large jump in latency. For each order of magnitude increase, we
add a + sign, up to 3. For order of magnitude decreases, we add a -
sign, up to 3.

We also needed to add a basic ring buffer of integers and a simple Mean
calculation.

We no longer generate i386 linux binaries for release. We got a little
ahead of ourselves with that one.